requirements.txt
Django==3.2
djangorestframework==3.14.0
requests==2.28.1
celery==5.2.7
transitions==0.9.0
python-decouple==3.8
tenacity==8.2.2